Add fragments demo to the showcase
